backend/requirements.txt
# VOXIS 4 Dense — Backend Requirements
# Powered by Trinity 8.1 | Built by Glass Stone

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
gunicorn>=21.2.0

# Audio Processing Core
numpy>=1.22
scipy>=1.11.0
soundfile>=0.12.1
librosa>=0.10.0
pydub>=0.25.1
resampy>=0.4.2

# Noise Reduction
noisereduce>=3.0.0

# Deep Filtering (Rikorose/DeepFilterNet)
deepfilternet>=0.5.6

# UVR5 Source Separation (Anjok07/ultimatevocalremovergui)
audio-separator[cpu]>=0.17.0

# PyTorch (CPU default — GPU: see setup.sh)
torch>=2.0.0
torchaudio>=2.0.0

# Utilities
werkzeug>=3.0.0
